Revert "feat: run extra ingest plugins from a trusted external Python (#21)"
This reverts commit 0fd6709ac2aa14e575f28ca354ab911e08a4e895.
