+++ docs/python-api.md
@@ -219,7 +219,6 @@ convert(
     "input.pdf",
     "output.vera",
     model="hashing",
-    parser="pymupdf",
     chunk_size=500,
     overlap=75,
     store_original=True,
@@ -228,9 +227,7 @@ convert(
 ```
 
 `vera-ingest` parses and chunks the source, creates `ChunkRecord` objects and
-optional attachments, then writes them through `VeraDocument`. Extra parsers
-register under `vera.ingest_pipelines`; see
-[Creating an ingest pipeline plugin](creating-an-ingest-pipeline.md).
+optional attachments, then writes them through `VeraDocument`.
 
 ## Corpus and library indexes
