+++ mkdocs.yml
@@ -128,7 +128,6 @@ nav:
       - Overview and concepts: packages/vera-ingest.md
       - Guides:
           - Convert documents: conversion.md
-          - Creating an ingest pipeline plugin: creating-an-ingest-pipeline.md
       - Examples:
           - Python examples: packages/vera-ingest-examples.md
       - API reference:
