+++ packages/vera-app/README.md
@@ -1,8 +1,8 @@
 # vera-app
 
-`vera-app` contains the Electron/React desktop application, its Python
-sidecar, and the shipped `vera_plugin_host` worker used with a user-selected
-interpreter for optional ingest plugins.
+`vera-app` contains the Electron/React desktop application and its Python
+sidecar. It composes `vera-doc` for storage/search with `vera-ingest` for
+conversion.
 
 See the [vera-app documentation](https://dkylewillis.github.io/vera/packages/vera-app/)
 for installation, user workflows, and architecture.
