+++ packages/vera-app/src/renderer/types.ts
@@ -2,9 +2,6 @@ import type {
   AppSettings,
   ChatCitationResult,
   CredentialResult,
-  ExternalPythonConfig,
-  PipelineDescriptor,
-  PythonEnvironmentProbe,
   Session,
   StreamEvent,
 } from '../shared/contracts';
@@ -16,11 +13,8 @@ export type {
   ContentPart,
   ContextChunkResult,
   CredentialResult,
-  ExternalPythonConfig,
   FigureResult,
-  PipelineDescriptor,
   ProviderProfile,
-  PythonEnvironmentProbe,
   RegionResult,
   SearchResult,
   Session,
@@ -49,9 +43,6 @@ export interface VeraApi {
   skipConversion(requestId: string): Promise<{ skipped: boolean }>;
   getSettings(): Promise<AppSettings>;
   saveSettings(settings: AppSettings): Promise<AppSettings>;
-  pickPythonInterpreter(): Promise<string | null>;
-  validatePythonEnvironment(executable: string, artifactsPath?: string): Promise<PythonEnvironmentProbe>;
-  refreshExternalPipelines(): Promise<PythonEnvironmentProbe>;
   saveApiKey(providerId: string, apiKey: string): Promise<CredentialResult>;
   clearApiKey(providerId: string): Promise<CredentialResult>;
   saveHfToken(token: string): Promise<CredentialResult>;
