--- packages/vera-app/src/vera_plugin_host/__main__.py
@@ -1,4 +1,0 @@
-from .worker import main
-
-if __name__ == "__main__":
-    raise SystemExit(main())
