--- tests/support/echo_ingest_plugin/pyproject.toml
@@ -1,19 +1,0 @@
-[project]
-name = "echo-ingest-plugin"
-version = "0.0.1"
-description = "Test ingest pipeline used by VERA plugin-host tests"
-requires-python = ">=3.10"
-dependencies = ["vera-ingest"]
-
-[project.entry-points."vera.ingest_pipelines"]
-echo = "echo_ingest_plugin:create_pipeline"
-
-[project.entry-points."vera.ingest_pipeline_descriptors"]
-echo = "echo_ingest_plugin:create_descriptor"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/echo_ingest_plugin"]
